+++ pkgs/jnigen/lib/src/bindings/renamer.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../config/config.dart';
+import '../config/config_types.dart';
 import '../elements/elements.dart';
 import '../logging/logging.dart';
 import 'visitor.dart';