+++ pkgs/ffigen/test/native_objc_test/category_test.dart
@@ -17,10 +17,6 @@ import 'util.dart';
 
 void main() {
   group('categories', () {
-    setUpAll(() {
-      loadLibrary();
-    });
-
     test('Category methods', () {
       final thing = Thing();
       expect(thing.add(1000, Y: 234), 1234);