+++ pkgs/ffigen/test/native_objc_test/global_native_test.dart
@@ -17,10 +17,6 @@ import 'util.dart';
 
 void main() {
   group('global using @Native', () {
-    setUpAll(() {
-      loadLibrary();
-    });
-
     test('Global string', () {
       expect(globalString.toDartString(), 'Hello World');
       globalString = 'Something else'.toNSString();