+++ pkgs/ffigen/test/native_objc_test/isolate_test.dart
@@ -19,13 +19,8 @@ import 'util.dart';
 
 void main() {
   group('isolate', () {
-    setUpAll(() {
-      loadLibrary();
-    });
-
     // Runs on other isolate (can't use expect function).
     void sendingObjectTest(SendPort sendPort) async {
-      loadLibrary();
       final port = ReceivePort();
       final queue = StreamQueue(port);
       sendPort.send(port.sendPort);
@@ -73,7 +68,6 @@ void main() {
       sendable.value = 123;
 
       final oldValue = await Isolate.run(() {
-        loadLibrary();
         final oldValue = sendable!.value;
         sendable!.value = 456;
         return oldValue;
@@ -92,7 +86,6 @@ void main() {
 
     // Runs on other isolate (can't use expect function).
     void sendingBlockTest(SendPort sendPort) async {
-      loadLibrary();
       final port = ReceivePort();
       final queue = StreamQueue(port);
       sendPort.send(port.sendPort);
@@ -148,7 +141,6 @@ void main() {
       ObjCBlock<Void Function(Int32)>? block = makeBlock(completer);
 
       await Isolate.run(() {
-        loadLibrary();
         block!(123);
       });
       final value = await completer.future;
@@ -170,7 +162,6 @@ void main() {
       expect(sendable.ref.isReleased, isFalse);
 
       final (oldIsReleased, newIsReleased) = await Isolate.run(() {
-        loadLibrary();
         final oldIsReleased = sendable.ref.isReleased;
         sendable!.ref.release();
         return (oldIsReleased, sendable.ref.isReleased);
@@ -191,7 +182,6 @@ void main() {
       expect(sendable.ref.isReleased, isFalse);
 
       await Isolate.run(() {
-        loadLibrary();
         sendable!.ref.release();
       });