+++ pkgs/ffigen/test/native_objc_test/log_test.dart
@@ -17,10 +17,6 @@ import 'util.dart';
 
 void main() {
   group('log_test', () {
-    setUpAll(() {
-      loadLibrary();
-    });
-
     test('Duplicate method log spam', () {
       final logs = <String>[];
       final logger = createTestLogger(