+++ pkgs/ffigen/test/native_objc_test/runtime_version_test.dart
@@ -17,10 +17,6 @@ import 'util.dart';
 
 void main() {
   group('runtime version check', () {
-    setUpAll(() {
-      loadLibrary();
-    });
-
     test('Interface', () {
       expect(() => FutureAPIInterface(), throwsA(isA<OsVersionError>()));
     });